+++ packages/website/src/pages/character/[id]/index.tsx
@@ -2,8 +2,8 @@ import { Divider, Image, Menu, MenuProps, Typography } from '@bangumi/design'
 import React from 'react'
 import { useParams } from 'react-router-dom'
 import InfoBoxComp from '../../../components/InfoBox'
-import { useCharacter } from '../../../hooks/useCharacter'
-import { useRelatedSubjectsOfCharacters } from '../../../hooks/useRelatedSubjects'
+import { useCharacter } from '../../../hooks/use-character'
+import { useRelatedSubjectsOfCharacters } from '../../../hooks/use-related-subjects'
 import RelatedSubject from './components/RelatedSubject'
 import styles from './index.module.less'
